+++ package.json
@@ -9,7 +9,7 @@
     "axios": "^0.19.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-router": "^5.1.2",
+    "react-router": "^3.2.6",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.1"
   },
